refactor and dashboard with stats

+++ src/redux/basketReducer/basketThunk.tsx
@@ -5,21 +5,24 @@ import { AppStateType } from '../redux-store'
 import * as AC from './basketActionCreator'
 
 export const deleteProductFromBasket = (productId: string) => {
-	return async (dispatch: Dispatch) => {
+	return async (dispatch: Dispatch): Promise<void> => {
 		dispatch(AC.removeFromBasket(productId))
 		dispatch(AC.counterSumBasket())
 	}
 }
 
 export const removeCountProduct = (productId: string) => {
-	return async (dispatch: Dispatch, getState: any) => {
+	return async (
+		dispatch: Dispatch,
+		getState: () => AppStateType,
+	): Promise<void> => {
 		const { basket } = getState()
 
 		const productInBasket = basket.productsBasket.find(
 			(productBasket: IProductBasket) => productBasket.id === productId,
 		)
 
-		if (productInBasket.count === 1) {
+		if (productInBasket && productInBasket.count === 1) {
 			dispatch(AC.removeFromBasket(productId))
 		} else {
 			dispatch(AC.decrementCount(productId))
@@ -29,7 +32,10 @@ export const removeCountProduct = (productId: string) => {
 }
 
 export const addToBasket = (productId: string) => {
-	return async (dispatch: Dispatch, getState: () => AppStateType) => {
+	return async (
+		dispatch: Dispatch,
+		getState: () => AppStateType,
+	): Promise<void> => {
 		const { basket } = getState()
 
 		let productFromDB = await productService.getProduct(productId)
